Fix: MAS was removed as customizer, I re-added her.
Note: to my knowledge, MAS was on her way to development on
SMF codebase. I consider that for 2.1-dev, acknowledgment
should be given for customization, and code contribution.
I expect that before anyone even thinks to a beta, the intention
to acknowledge community contributions is a given.

I will let her and the community use their judgment and state
clearly if she considers otherwise. (same as for everyone else).

No one can really make this call for SMF development, and no one
can or should handle it privately any longer, so I offer you my own
perspective, at this time.

Signed-off-by: Norv <a.w.norv@gmail.com>