Added Labradoodle as customizer. Labradoodle has my thanks for his intentions to pick up development in SM/SMF project. Contributions may do justice to these intentions in the future. At this time, I consider appropriate to acknowledge his contributions to customization on the credits page in the software, and the code contributions to the core in the (upcoming) contributors file.
It is up to him and the development community around SM/SMF to state
clearly if they think otherwise.

No one can possibly make any "call" in the team of SM/SMF on matters
like this: and neither do I. I offer you my perspective, on adjustments
to the current 2.1-dev credits page.

Signed-off-by: Norv <a.w.norv@gmail.com>